Split end to end tests by common test cases
End to end tests have been split into the following files. Tests have
been divided into the following categories.
- Filtered  (include/excludes)
- Incremental
- Plugin
- Special characters
- Everything else (original file)

Sql script files moved into resources folder

Authored-by: Kevin Yeap <kyeap@pivotal.io>
