+++ end_to_end/filtered_test.go
@@ -1,0 +1,326 @@
+package end_to_end_test
+
+import (
+	"os"
+
+	"github.com/greenplum-db/gp-common-go-libs/iohelper"
+	"github.com/greenplum-db/gp-common-go-libs/testhelper"
+	"github.com/greenplum-db/gpbackup/utils"
+	. "github.com/onsi/ginkgo"
+)
+
+var _ = Describe("End to End Filtered tests", func() {
+	BeforeEach(func() {
+		end_to_end_setup()
+	})
+	AfterEach(func() {
+		end_to_end_teardown()
+	})
+
+	Describe("Backup include filtering", func() {
+		It("runs gpbackup and gprestore with include-schema backup flag and compression level", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--include-schema", "public",
+				"--compression-level", "2")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 20)
+			assertDataRestored(restoreConn, publicSchemaTupleCounts)
+			assertArtifactsCleaned(restoreConn, timestamp)
+		})
+		It("runs gpbackup and gprestore with include-table backup flag", func() {
+			skipIfOldBackupVersionBefore("1.4.0")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--include-table", "public.foo",
+				"--include-table", "public.sales",
+				"--include-table", "public.myseq1",
+				"--include-table", "public.myview1")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 16)
+			assertDataRestored(restoreConn, map[string]int{"public.foo": 40000})
+		})
+		It("runs gpbackup and gprestore with include-table-file backup flag", func() {
+			skipIfOldBackupVersionBefore("1.4.0")
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/include-tables.txt")
+			utils.MustPrintln(includeFile, "public.sales\npublic.foo\npublic.myseq1\npublic.myview1")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--include-table-file", "/tmp/include-tables.txt")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 16)
+			assertDataRestored(restoreConn, map[string]int{"public.sales": 13, "public.foo": 40000})
+
+			_ = os.Remove("/tmp/include-tables.txt")
+		})
+		It("runs gpbackup and gprestore with include-schema-file backup flag", func() {
+			skipIfOldBackupVersionBefore("1.17.0")
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/include-schema.txt")
+			utils.MustPrintln(includeFile, "public")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--include-schema-file", "/tmp/include-schema.txt")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 20)
+
+			_ = os.Remove("/tmp/include-schema.txt")
+		})
+		It("runs gpbackup with --include-table flag with partitions (non-special chars)", func() {
+			testhelper.AssertQueryRuns(backupConn,
+				`CREATE TABLE public.testparent (id int, rank int, year int, gender
+char(1), count int )
+DISTRIBUTED BY (id)
+PARTITION BY LIST (gender)
+( PARTITION girls VALUES ('F'),
+  PARTITION boys VALUES ('M'),
+  DEFAULT PARTITION other );
+			`)
+			defer testhelper.AssertQueryRuns(backupConn,
+				`DROP TABLE public.testparent`)
+
+			testhelper.AssertQueryRuns(backupConn,
+				`insert into public.testparent values (1,1,1,'M',1)`)
+			testhelper.AssertQueryRuns(backupConn,
+				`insert into public.testparent values (0,0,0,'F',1)`)
+
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir,
+				"--include-table", `public.testparent_1_prt_girls`,
+				"--leaf-partition-data")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir)
+
+			assertRelationsCreated(restoreConn, 4)
+			localSchemaTupleCounts := map[string]int{
+				`public.testparent_1_prt_girls`: 1,
+				`public.testparent`:             1,
+			}
+			assertDataRestored(restoreConn, localSchemaTupleCounts)
+			assertArtifactsCleaned(restoreConn, timestamp)
+		})
+	})
+	Describe("Restore include filtering", func() {
+		It("runs gpbackup and gprestore with include-schema restore flag", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--include-schema", "schema2")
+
+			assertRelationsCreated(restoreConn, 17)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+		})
+		It("runs gpbackup and gprestore with include-schema-file restore flag", func() {
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/include-schema.txt")
+			utils.MustPrintln(includeFile, "schema2")
+			utils.MustPrintln(includeFile, "public")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--include-schema-file", "/tmp/include-schema.txt")
+
+			assertRelationsCreated(restoreConn, 37)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+		})
+		It("runs gpbackup and gprestore with include-table restore flag", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--include-table", "public.foo",
+				"--include-table", "public.sales",
+				"--include-table", "public.myseq1",
+				"--include-table", "public.myview1")
+
+			assertRelationsCreated(restoreConn, 16)
+			assertDataRestored(restoreConn, map[string]int{
+				"public.sales": 13, "public.foo": 40000})
+		})
+		It("runs gpbackup and gprestore with include-table-file restore flag", func() {
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/include-tables.txt")
+			utils.MustPrintln(includeFile,
+				"public.sales\npublic.foo\npublic.myseq1\npublic.myview1")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--include-table-file", "/tmp/include-tables.txt")
+
+			assertRelationsCreated(restoreConn, 16)
+			assertDataRestored(restoreConn, map[string]int{
+				"public.sales": 13, "public.foo": 40000})
+
+			_ = os.Remove("/tmp/include-tables.txt")
+		})
+		It("runs gpbackup and gprestore with include-table restore flag against a leaf partition", func() {
+			skipIfOldBackupVersionBefore("1.7.2")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--leaf-partition-data")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--include-table", "public.sales_1_prt_jan17")
+
+			assertRelationsCreated(restoreConn, 13)
+			assertDataRestored(restoreConn, map[string]int{
+				"public.sales": 1, "public.sales_1_prt_jan17": 1})
+		})
+		It("runs gpbackup and gprestore with include-table restore flag which implicitly filters schema restore list", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--include-table", "schema2.foo3")
+			assertRelationsCreated(restoreConn, 1)
+			assertDataRestored(restoreConn, map[string]int{"schema2.foo3": 100})
+		})
+		It("runs gprestore with --include-table flag to only restore tables specified ", func() {
+			testhelper.AssertQueryRuns(backupConn,
+				"CREATE TABLE public.table_to_include_with_stats(i int)")
+			defer testhelper.AssertQueryRuns(backupConn,
+				"DROP TABLE public.table_to_include_with_stats")
+			testhelper.AssertQueryRuns(backupConn,
+				"INSERT INTO public.table_to_include_with_stats VALUES (1)")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--include-table", "public.table_to_include_with_stats")
+
+			assertRelationsCreated(restoreConn, 1)
+
+			localSchemaTupleCounts := map[string]int{
+				`public."table_to_include_with_stats"`: 1,
+			}
+			assertDataRestored(restoreConn, localSchemaTupleCounts)
+			assertArtifactsCleaned(restoreConn, timestamp)
+		})
+
+	})
+	Describe("Backup exclude filtering", func() {
+		It("runs gpbackup and gprestore with exclude-schema backup flag", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--exclude-schema", "public")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 17)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+		})
+		It("runs gpbackup and gprestore with exclude-schema-file backup flag", func() {
+			skipIfOldBackupVersionBefore("1.17.0")
+			excludeFile := iohelper.MustOpenFileForWriting("/tmp/exclude-schema.txt")
+			utils.MustPrintln(excludeFile, "public")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--exclude-schema-file", "/tmp/exclude-schema.txt")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 17)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+
+			_ = os.Remove("/tmp/exclude-schema.txt")
+		})
+		It("runs gpbackup and gprestore with exclude-table backup flag", func() {
+			skipIfOldBackupVersionBefore("1.4.0")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--exclude-table", "schema2.foo2",
+				"--exclude-table", "schema2.returns",
+				"--exclude-table", "public.myseq2",
+				"--exclude-table", "public.myview2")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, TOTAL_RELATIONS_AFTER_EXCLUDE)
+			assertDataRestored(restoreConn, map[string]int{
+				"schema2.foo3": 100,
+				"public.foo":   40000,
+				"public.holds": 50000,
+				"public.sales": 13})
+		})
+		It("runs gpbackup and gprestore with exclude-table-file backup flag", func() {
+			skipIfOldBackupVersionBefore("1.4.0")
+			excludeFile := iohelper.MustOpenFileForWriting("/tmp/exclude-tables.txt")
+			utils.MustPrintln(excludeFile,
+				"schema2.foo2\nschema2.returns\npublic.sales\npublic.myseq2\npublic.myview2")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--exclude-table-file", "/tmp/exclude-tables.txt")
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb")
+
+			assertRelationsCreated(restoreConn, 8)
+			assertDataRestored(restoreConn, map[string]int{
+				"schema2.foo3": 100,
+				"public.foo":   40000,
+				"public.holds": 50000})
+
+			_ = os.Remove("/tmp/exclude-tables.txt")
+		})
+	})
+	Describe("Restore exclude filtering", func() {
+		It("runs gpbackup and gprestore with exclude-schema restore flag", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--exclude-schema", "public")
+
+			assertRelationsCreated(restoreConn, 17)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+		})
+		It("runs gpbackup and gprestore with exclude-schema-file restore flag", func() {
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/exclude-schema.txt")
+			utils.MustPrintln(includeFile, "public")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--backup-dir", backupDir,
+				"--redirect-db", "restoredb",
+				"--exclude-schema-file", "/tmp/exclude-schema.txt")
+
+			assertRelationsCreated(restoreConn, 17)
+			assertDataRestored(restoreConn, schema2TupleCounts)
+
+			_ = os.Remove("/tmp/exclude-schema.txt")
+		})
+		It("runs gpbackup and gprestore with exclude-table restore flag", func() {
+			timestamp := gpbackup(gpbackupPath, backupHelperPath)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--exclude-table", "schema2.foo2",
+				"--exclude-table", "schema2.returns",
+				"--exclude-table", "public.myseq2",
+				"--exclude-table", "public.myview2")
+
+			assertRelationsCreated(restoreConn, TOTAL_RELATIONS_AFTER_EXCLUDE)
+			assertDataRestored(restoreConn, map[string]int{
+				"schema2.foo3": 100, "public.foo": 40000, "public.holds": 50000, "public.sales": 13})
+		})
+		It("runs gpbackup and gprestore with exclude-table-file restore flag", func() {
+			includeFile := iohelper.MustOpenFileForWriting("/tmp/exclude-tables.txt")
+			utils.MustPrintln(includeFile,
+				"schema2.foo2\nschema2.returns\npublic.myseq2\npublic.myview2")
+			timestamp := gpbackup(gpbackupPath, backupHelperPath,
+				"--backup-dir", backupDir)
+			gprestore(gprestorePath, restoreHelperPath, timestamp,
+				"--redirect-db", "restoredb",
+				"--backup-dir", backupDir,
+				"--exclude-table-file", "/tmp/exclude-tables.txt")
+
+			assertRelationsCreated(restoreConn, TOTAL_RELATIONS_AFTER_EXCLUDE)
+			assertDataRestored(restoreConn, map[string]int{
+				"public.sales": 13,
+				"public.foo":   40000})
+
+			_ = os.Remove("/tmp/exclude-tables.txt")
+		})
+	})
+})
