+++ end_to_end/resources/gpdb4_objects.sql
