+++ end_to_end/resources/gpdb6_objects.sql
