+++ end_to_end/resources/test_tables_ddl.sql
